Refactor helper utilities into focused classes
