+++ src/com/intellij/advancedExpressionFolding/expression/math/advanced/Exp.kt
@@ -3,7 +3,7 @@ package com.intellij.advancedExpressionFolding.expression.math.advanced
 import com.intellij.advancedExpressionFolding.expression.Expression
 import com.intellij.advancedExpressionFolding.expression.Function
 import com.intellij.advancedExpressionFolding.expression.math.ArithmeticExpression
-import com.intellij.advancedExpressionFolding.processor.util.Helper
+import com.intellij.advancedExpressionFolding.processor.util.SuperscriptUtil
 import com.intellij.lang.folding.FoldingDescriptor
 import com.intellij.openapi.editor.Document
 import com.intellij.openapi.editor.FoldingGroup
@@ -17,7 +17,7 @@ class Exp(
     operands: List<Expression>
 ) : Function(element, textRange, "exp", operands), ArithmeticExpression {
     override fun supportsFoldRegions(document: Document, parent: Expression?): Boolean {
-        return Helper.superscript(operands[0].element.text) != null
+        return SuperscriptUtil.superscript(operands[0].element.text) != null
     }
 
     override fun buildFoldRegions(
@@ -32,7 +32,7 @@ class Exp(
                 element.node,
                 TextRange.create(textRange),
                 group,
-                "\uD835\uDC52" + Helper.superscript(operands[0].element.text)
+                "\uD835\uDC52" + SuperscriptUtil.superscript(operands[0].element.text)
             )
         )
         return descriptors.toTypedArray()
