+++ src/com/intellij/advancedExpressionFolding/processor/expression/PsiArrayAccessExpressionExt.kt
@@ -5,7 +5,7 @@ import com.intellij.advancedExpressionFolding.expression.literal.NumberLiteral
 import com.intellij.advancedExpressionFolding.expression.operation.collection.ArrayGet
 import com.intellij.advancedExpressionFolding.processor.core.BaseExtension
 import com.intellij.advancedExpressionFolding.processor.core.BuildExpressionExt
-import com.intellij.advancedExpressionFolding.processor.util.Helper
+import com.intellij.advancedExpressionFolding.processor.util.PsiTreeUtilEx
 import com.intellij.advancedExpressionFolding.settings.AdvancedExpressionFoldingSettings
 import com.intellij.openapi.editor.Document
 import com.intellij.psi.PsiArrayAccessExpression
@@ -23,7 +23,7 @@ object PsiArrayAccessExpressionExt : BaseExtension() {
         if (!isLeftSideAssignment && index != null && settings.state.getExpressionsCollapse) {
             val arrayExpression = BuildExpressionExt.getAnyExpression(element.arrayExpression, document)
             if (index is PsiBinaryExpression) {
-                val position: NumberLiteral? = Helper.getSlicePosition(element, arrayExpression, index, document)
+                val position: NumberLiteral? = PsiTreeUtilEx.getSlicePosition(element, arrayExpression, index, document)
                 if (position != null && position.number == -1) {
                     return ArrayGet(element, element.textRange, arrayExpression)
                 }
