+++ src/com/intellij/advancedExpressionFolding/processor/methodcall/optional/OptionalOfNullableMethodCall.kt
@@ -3,7 +3,7 @@ package com.intellij.advancedExpressionFolding.processor.methodcall.optional
 import com.intellij.advancedExpressionFolding.expression.Expression
 import com.intellij.advancedExpressionFolding.expression.operation.optional.OptionalOfNullable
 import com.intellij.advancedExpressionFolding.processor.methodcall.Context
-import com.intellij.advancedExpressionFolding.processor.util.Helper
+import com.intellij.advancedExpressionFolding.processor.util.MethodNameUtil
 import com.intellij.psi.PsiExpression
 import com.intellij.psi.PsiMethodCallExpression
 
@@ -16,7 +16,7 @@ class OptionalOfNullableMethodCall : AbstractOptionalMethodCall() {
         argument: PsiExpression,
         argumentExpression: Expression
     ): Expression? {
-        if (Helper.hasOptionalChainOperations(element)) {
+        if (MethodNameUtil.hasOptionalChainOperations(element)) {
             return OptionalOfNullable(
                 element,
                 element.textRange,
