+++ src/com/intellij/advancedExpressionFolding/processor/methodcall/optional/OptionalOrElseMethodCall.kt
@@ -3,7 +3,7 @@ package com.intellij.advancedExpressionFolding.processor.methodcall.optional
 import com.intellij.advancedExpressionFolding.expression.Expression
 import com.intellij.advancedExpressionFolding.expression.operation.optional.OptionalOrElseElvis
 import com.intellij.advancedExpressionFolding.processor.methodcall.Context
-import com.intellij.advancedExpressionFolding.processor.util.Helper
+import com.intellij.advancedExpressionFolding.processor.util.PsiTreeUtilEx
 import com.intellij.psi.PsiExpression
 import com.intellij.psi.PsiExpressionList
 import com.intellij.psi.PsiMethodCallExpression
@@ -17,7 +17,7 @@ class OptionalOrElseMethodCall : AbstractOptionalMethodCall() {
         argument: PsiExpression,
         argumentExpression: Expression
     ): Expression? {
-        if (Helper.findChildByTypeHierarchy(element, PsiExpressionList::class.java, PsiExpressionList::class.java) != null) {
+        if (PsiTreeUtilEx.findChildByTypeHierarchy(element, PsiExpressionList::class.java, PsiExpressionList::class.java) != null) {
             return OptionalOrElseElvis(
                 element,
                 element.textRange,
