+++ src/com/intellij/advancedExpressionFolding/processor/reference/ReferenceExpressionExt.kt
@@ -6,7 +6,7 @@ import com.intellij.advancedExpressionFolding.expression.operation.basic.Variabl
 import com.intellij.advancedExpressionFolding.processor.expression.AssignmentExpressionExt
 import com.intellij.advancedExpressionFolding.processor.language.ExperimentalExt
 import com.intellij.advancedExpressionFolding.processor.util.Consts
-import com.intellij.advancedExpressionFolding.processor.util.Helper
+import com.intellij.advancedExpressionFolding.processor.util.TypeUtil
 import com.intellij.psi.PsiIdentifier
 import com.intellij.psi.PsiMethodReferenceExpression
 import com.intellij.psi.PsiReferenceExpression
@@ -17,10 +17,10 @@ object ReferenceExpressionExt {
         val identifier = element.children.firstOrNull { it is PsiIdentifier } as? PsiIdentifier ?: return null
         val constant = Consts.SUPPORTED_CONSTANTS[identifier.text]
         if (constant != null) {
-            if (Helper.isSupportedClass(element) && constant is Number) {
+            if (TypeUtil.isSupportedClass(element) && constant is Number) {
                 return NumberLiteral(element, element.textRange, element.textRange, constant, true)
             }
-            if (Helper.isSupportedClass(element) && constant is String) {
+            if (TypeUtil.isSupportedClass(element) && constant is String) {
                 return Variable(element, element.textRange, null, constant, copy)
             }
         } else {
